added output for cleaned table
NB removed duplicated comparison, using the first occurrence of the comparison. But because it uses a certain number of sketches, the ANI value measured is not exactly the same in some cases. 
We could eg. use that to see if the variation is large as an indicator that the number of skeches should be increased. But still we will expect variation ...